fix api setting + validate

+++ BackEnd/src/main/java/com/fpt/capstone/backend/api/BackEnd/service/impl/AuditorAwareImpl.java
@@ -11,13 +11,13 @@ import org.springframework.stereotype.Component;
 import java.util.Optional;
 
 @Component
-public class AuditorAwareImpl implements AuditorAware<Long> {
+public class AuditorAwareImpl implements AuditorAware<Integer> {
 
     @Autowired
     UserRepository userRepository;
 
     @Override
-    public Optional<Long> getCurrentAuditor() {
+    public Optional<Integer> getCurrentAuditor() {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
         //CustomUserDetails userDetails = (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication()
         //        .getPrincipal();
